add vector swizzle
