+++ python/taichi/math/__init__.py
@@ -1,3 +1,9 @@
+"""Taichi math module.
+
+The math module supports glsl-style vectors, matrices and functions.
+"""
 from taichi.math.mathimpl import *
+from taichi.math.vectypes import *
 
 del mathimpl
+del vectypes
