+++ tests/python/test_math_vectors.py
@@ -4,23 +4,27 @@ from tests import test_utils
 
 @test_utils.test()
 def test_vector_swizzle_python():
-    a = ti.vec3(1, 2, 3)
+    a = ti.math.vec3(1, 2, 3)
     assert all(a.gbr == (2, 3, 1))
     a.bgr = (2, 3, 3)
     assert all(a.rgb == (3, 3, 2))
-    a.r = 1
-    assert a.r == 1
+    b = ti.math.vec3(1)
+    assert all((a + b).rgb == (4, 4, 3))
+    M = ti.Matrix([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+    assert all((M @ b) == (1, 1, 1))
 
 
 @test_utils.test(debug=True)
 def test_vector_swizzle_taichi():
     @ti.kernel
     def foo():
-        a = ti.vec3(1, 2, 3)
+        a = ti.math.vec3(1, 2, 3)
         assert all(a.gbr == (2, 3, 1))
         a.bgr = (2, 3, 3)
         assert all(a.rgb == (3, 3, 2))
-        a.r = 1
-        assert a.r == 1
-
+        b = ti.math.vec3(1)
+        assert all((a + b).rgb == (4, 4, 3))
+        M = ti.Matrix([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+        assert all((M @ b) == (1, 1, 1))
+    
     foo()
